refactor:moved button logic + button to its own component
