+++ packages/widget/src/components/SendToWalletConfirmButton/index.ts
@@ -1,0 +1,1 @@
+export * from './SendToWalletConfirmButton';
